# Accelerate STM32 QSPI read/write operations by DMA