--- a/storage/filesystem/fat/source/FATFileSystem.cpp
+++ b/storage/filesystem/fat/source/FATFileSystem.cpp
@@ -155,7 +155,11 @@ extern "C" DWORD get_fattime(void)
 
 extern "C" void *ff_memalloc(UINT size)
 {
+#if defined (__DCACHE_PRESENT) && (__DCACHE_PRESENT == 1U)
+    return aligned_alloc(32U, size);
+#else
     return malloc(size);
+#endif
 }
 
 extern "C" void ff_memfree(void *p)