--- a/targets/TARGET_STM/TARGET_STM32F7/objects.h
+++ b/targets/TARGET_STM/TARGET_STM32F7/objects.h
@@ -43,8 +43,10 @@
 #include "stm32f7xx_ll_rcc.h"
 
 #include "stm_dma_info.h"
+#if MBED_CONF_RTOS_PRESENT
 #include "cmsis_os.h"
 #include "cmsis_os2.h"
+#endif
 
 #ifdef __cplusplus
 extern "C" {
@@ -129,8 +131,10 @@ struct qspi_s {
     PinName sclk;
     PinName ssel;
     bool dmaInitialized;
+#if MBED_CONF_RTOS_PRESENT
     osSemaphoreId_t semaphoreId;
     osRtxSemaphore_t semaphoreMem;
+#endif
 };
 #endif