--- a/targets/TARGET_STM/TARGET_STM32G4/objects.h
+++ b/targets/TARGET_STM/TARGET_STM32G4/objects.h
@@ -27,8 +27,10 @@
 #include "stm32g4xx_ll_rcc.h"
 
 #include "stm_dma_info.h"
+#if MBED_CONF_RTOS_PRESENT
 #include "cmsis_os.h"
 #include "cmsis_os2.h"
+#endif
 
 #ifdef __cplusplus
 extern "C" {
@@ -111,8 +113,10 @@ struct qspi_s {
     PinName sclk;
     PinName ssel;
     bool dmaInitialized;
+#if MBED_CONF_RTOS_PRESENT
     osSemaphoreId_t semaphoreId;
     osRtxSemaphore_t semaphoreMem;
+#endif
 };
 #endif