--- a/targets/TARGET_STM/TARGET_STM32H7/stm_dma_info.h
+++ b/targets/TARGET_STM/TARGET_STM32H7/stm_dma_info.h
@@ -45,12 +45,12 @@ static const DMALinkInfo SPIRxDMALinks[] = {
 /// Mapping from QSPI/OSPI index to DMA link info
 #if defined(OCTOSPI1)
 static const DMALinkInfo OSPIDMALinks[] = {
-        {4, 0, MDMA_REQUEST_OCTOSPI1_FIFO_TH},
-        {4, 1, MDMA_REQUEST_OCTOSPI2_FIFO_TH}
+        {MDMA_IDX, 0, MDMA_REQUEST_OCTOSPI1_FIFO_TH},
+        {MDMA_IDX, 1, MDMA_REQUEST_OCTOSPI2_FIFO_TH}
 };
 #else
 static const DMALinkInfo QSPIDMALinks[] = {
-        {4, 0, MDMA_REQUEST_QUADSPI_FIFO_TH},
+        {MDMA_IDX, 0, MDMA_REQUEST_QUADSPI_FIFO_TH},
 };
 #endif