--- a/targets/TARGET_STM/TARGET_STM32L4/objects.h
+++ b/targets/TARGET_STM/TARGET_STM32L4/objects.h
@@ -32,6 +32,8 @@
 #include "stm32l4xx_ll_rcc.h"
 
 #include "stm_dma_utils.h"
+#include "cmsis_os.h"
+#include "cmsis_os2.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -120,12 +122,15 @@ struct qspi_s {
     PinName sclk;
     PinName ssel;
     bool dmaInitialized;
+    osSemaphoreId_t semaphoreId;
+    osRtxSemaphore_t semaphoreMem;
 };
 #endif
 
 #if DEVICE_OSPI
 struct ospi_s {
     OSPI_HandleTypeDef handle;
+    IRQn_Type ospiIRQ;
     OSPIName ospi;
     PinName io0;
     PinName io1;
@@ -138,8 +143,9 @@ struct ospi_s {
     PinName sclk;
     PinName ssel;
     PinName dqs;
-    IRQn_Type ospiIRQ;
-    bool dmaInitialized;    
+    bool dmaInitialized;
+    osSemaphoreId_t semaphoreId;
+    osRtxSemaphore_t semaphoreMem;
 };
 #endif