--- a/targets/TARGET_STM/TARGET_STM32L5/stm_dma_info.h
+++ b/targets/TARGET_STM/TARGET_STM32L5/stm_dma_info.h
@@ -38,5 +38,9 @@ static const DMALinkInfo SPIRxDMALinks[] = {
         {1, 6, DMA_REQUEST_SPI3_RX},
 };
 
+/// Mapping from OSPI index to DMA link info
+static const DMALinkInfo OSPIDMALinks[] = {
+        {1, 7, DMA_REQUEST_OCTOSPI1},
+};
 
 #endif //MBED_OS_STM_DMA_INFO_H