--- a/targets/TARGET_STM/TARGET_STM32U5/objects.h
+++ b/targets/TARGET_STM/TARGET_STM32U5/objects.h
@@ -27,6 +27,8 @@
 #include "stm32u5xx_ll_pwr.h"
 #include "stm32u5xx_ll_rcc.h"
 
+#include "stm_dma_info.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -112,6 +114,8 @@ struct qspi_s {
     PinName io3;
     PinName sclk;
     PinName ssel;
+    IRQn_Type qspiIRQ;
+    bool dmaInitialized;
 };
 
 struct ospi_s {
@@ -128,6 +132,8 @@ struct ospi_s {
     PinName sclk;
     PinName ssel;
     PinName dqs;
+    IRQn_Type ospiIRQ;
+    bool dmaInitialized;
 };
 
 #ifdef __cplusplus