--- a/targets/TARGET_STM/TARGET_STM32U5/stm_dma_info.h
+++ b/targets/TARGET_STM/TARGET_STM32U5/stm_dma_info.h
@@ -41,7 +41,9 @@ static const DMALinkInfo SPIRxDMALinks[] = {
 /// Mapping from OSPI index to DMA link info
 static const DMALinkInfo OSPIDMALinks[] = {
         {1, 6, GPDMA1_REQUEST_OCTOSPI1},
+#if defined(OCTOSPI2)
         {1, 7, GPDMA1_REQUEST_OCTOSPI2}
+#endif
 };
 
 #endif //MBED_OS_STM_DMA_INFO_H