--- a/targets/TARGET_STM/TARGET_STM32WB/stm_dma_info.h
+++ b/targets/TARGET_STM/TARGET_STM32WB/stm_dma_info.h
@@ -40,5 +40,10 @@ static const DMALinkInfo SPIRxDMALinks[] = {
 #endif
 };
 
+static const DMALinkInfo QSPIDMALinks[] = {
+#if defined(QUADSPI)
+        {1, 5, DMA_REQUEST_QUADSPI},
+#endif
+};
 
 #endif //MBED_OS_STM_DMA_INFO_H