--- a/targets/TARGET_STM/stm_dma_ip_v1.h
+++ b/targets/TARGET_STM/stm_dma_ip_v1.h
@@ -32,6 +32,8 @@
 // For STM32H7, MDMA has 16 channels, while DMA1, DMA2 and BDMA all have 8 channels
 #define NUM_DMA_CONTROLLERS 5
 #define MAX_MDMA_CHANNELS 16
+#define BDMA_IDX 3
+#define MDMA_IDX 4
 #else
 #ifdef DMA1
 #ifdef DMA2